requirements.txt
httpx
python-dotenv
pytest
PyGithub
openai


